% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addTimestamp}
\alias{addTimestamp}
\title{Function that adds a timestamp to a data frame}
\usage{
addTimestamp(dataframe, frequency, start = attr(dataframe, which =
  "measurementStart"))
}
\arguments{
\item{dataframe}{The name of the dataframe.}

\item{frequency}{The frequency of measurement in Hz.}

\item{start}{The initial timestamp.}
}
\description{
Takes a data frame (that has the starting value of measurements as an attribute)
and sampling frequency as input. Outputs the data frame with additional column containing
timestamps of measurement.
}

