% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getStartofMeasurement}
\alias{getStartofMeasurement}
\title{Read date and time of the start of measurement}
\usage{
getStartofMeasurement(dirname)
}
\arguments{
\item{dirname}{The path to the directory containing the bat_adc.txt files.}
}
\description{
Uses the function readBatteryStatus() to get data and
extracts and returns the timestamp of the first entry, assuming this is the
 start of all the measurements.
}

