% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readAccel}
\alias{readAccel}
\title{Read acceleremeter data}
\usage{
readAccel(dirname)
}
\arguments{
\item{dirname}{The path to the directory containing the accel.bin file
(assumes bat_adc.txt is in the same directory).}
}
\description{
Reads the accel.bin file saved by the cortrium C3 and outputs
a dataframe with columns for each of the three accelerometer readings.
Sampled at 25 Hz.
}
\details{
Returns the start and end of measurement as an attribute measurementStart and
measurementEnd.
}

