% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readBatteryStatus}
\alias{readBatteryStatus}
\title{Read battery status data}
\usage{
readBatteryStatus(dirname)
}
\arguments{
\item{dirname}{The path to the directory containing the bat_adc.txt files.}
}
\description{
Read the records about battery status contained in the bat_adc.txt
file and output a data frame with column containing the timestamp,
 ??? and battery status.
}

