% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readECG}
\alias{readECG}
\title{Read ECG records}
\usage{
readECG(dirname)
}
\arguments{
\item{dirname}{The path to the directory containing the ecgX_raw.bin files.}
}
\description{
Read the ECG records contained in the ecgX_raw.bin files and
outputs a dataframe with columns for each of the three ECG leads.
Sampled at 250 Hz.

NOTE: This is raw ECG data. Nothing is currently done with values
greater than 32676 - in the matlab scripts they seem to be set to 0.
}

