% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readRespiration}
\alias{readRespiration}
\title{Read respiration rate records}
\usage{
readRespiration(dirname)
}
\arguments{
\item{dirname}{The path to the directory containing the resp_raw.bin files
(assumes bat_adc.txt is in the same directory).}
}
\description{
Read the respiration rate records contained in the resp_raw.bin
file and output them in a data frame. Sampled at 250 Hz.

NOTE: This is raw respiration rate data. It has to be further processed to
get a sensible respiration rate (local maxima imply a maximum chest expansion?).
}

