% Generated by roxygen2 (4.0.1): do not edit by hand
\name{beatsFromECG}
\alias{beatsFromECG}
\title{A simple function to extract beat timings from ECG data}
\usage{
beatsFromECG(dirname, cutoffValue = 200, highpassFreq = 0.3)
}
\arguments{
\item{dirname}{The path to the directory containing the ecgX_raw.bin files.}
}
\description{
Reads the ECG signal data and applies a highpass filter to
exagurate the spikes of the QRS complex. Than looks where the spikes are higher
than some value and assumes that represents the R part of the ECG. Looks
where the next peak is more than 25 values (100ms) away from the previous one
and assumes this as the start of the next heart beat. Uses the
ECG1 records. This is a simple method - better ones exist -
for a quick implementation.
Outputs data frame with columns representing beat times in seconds.
}

